All the files I have on the current version of the Beach Scene
I know there are about 5,000 or so files. I do not believe they are all necessary, but I didn't want to erase anything just in case it affected the project somehow. You should be able to open the project by going to the Assets folder > Scenes > Beach Scene. Please keep me posted if continues causing errors when trying to open it.